Document code and organize into files
