+++ lib/music_player/demixer/demixer.dart
@@ -1,63 +1,14 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:just_audio/just_audio.dart';
 import 'package:musbx/music_player/demixer/demixer_api.dart';
+import 'package:musbx/music_player/demixer/demixer_api_exceptions.dart';
+import 'package:musbx/music_player/demixer/stem.dart';
 import 'package:musbx/music_player/music_player.dart';
 import 'package:musbx/music_player/music_player_component.dart';
 import 'package:musbx/music_player/song.dart';
 import 'package:musbx/widgets.dart';
-
-final DemixerApi _api = DemixerApi();
-
-class Stem {
-  Stem(this.type) {
-    player.volumeStream.listen((value) => volumeNotifier.value = value);
-  }
-
-  final StemType type;
-
-  /// Whether this stem is enabled and should be played.
-  set enabled(bool value) {
-    if (value == false) player.pause();
-    if (value == true &&
-        MusicPlayer.instance.demixer.enabled &&
-        MusicPlayer.instance.isPlaying) player.play();
-    enabledNotifier.value = value;
-  }
-
-  bool get enabled => enabledNotifier.value;
-  final ValueNotifier<bool> enabledNotifier = ValueNotifier(true);
-
-  /// The volume this stem is played at. Must be between 0 and 1.
-  set volume(double value) => player.setVolume(value.clamp(0, 1));
-  double get volume => volumeNotifier.value;
-  final ValueNotifier<double> volumeNotifier = ValueNotifier(1.0);
-
-  final AndroidEqualizer equalizer = AndroidEqualizer();
-  late final AudioPlayer player = AudioPlayer(
-    audioPipeline: AudioPipeline(androidAudioEffects: [
-      if (Platform.isAndroid) equalizer,
-    ]),
-  );
-
-  /// Download and prepare [player] for playing this stem of [song].
-  Future<void> loadStemFile(String song) async {
-    File? file = await _api.downloadStem(song, type);
-    if (file == null) return;
-    await player.setAudioSource(AudioSource.file(file.path));
-  }
-}
-
-class StemsNotifier extends ValueNotifier<List<Stem>> {
-  StemsNotifier(super.value) {
-    for (Stem stem in value) {
-      stem.enabledNotifier.addListener(notifyListeners);
-      stem.volumeNotifier.addListener(notifyListeners);
-    }
-  }
-}
 
 enum DemixerState {
   /// Loading hasn't started. E.g. the user hasn't selected a song yet.
@@ -79,16 +30,21 @@ enum DemixerState {
   error,
 }
 
+/// A component for [MusicPlayer] that is used to separate a song into stems and change the volume of those individually.
 class Demixer extends MusicPlayerComponent {
-  final Stem drums = Stem(StemType.drums);
-  final Stem bass = Stem(StemType.bass);
-  final Stem vocals = Stem(StemType.vocals);
-  final Stem other = Stem(StemType.other);
+  /// The Demixer API used internally.
+  final DemixerApi api = DemixerApi();
 
+  /// The stems that songs are being separated into.
   List<Stem> get stems => stemsNotifier.value;
-  late final StemsNotifier stemsNotifier =
-      StemsNotifier([drums, bass, vocals, other]);
+  late final StemsNotifier stemsNotifier = StemsNotifier([
+    Stem(StemType.drums),
+    Stem(StemType.bass),
+    Stem(StemType.vocals),
+    Stem(StemType.other)
+  ]);
 
+  /// The state of the Demixer.
   DemixerState get state => loadingStateNotifier.value;
   final ValueNotifier<DemixerState> loadingStateNotifier =
       ValueNotifier(DemixerState.inactive);
@@ -98,6 +54,7 @@ class Demixer extends MusicPlayerComponent {
   /// If `true`, the current song has been separated and mixed, and the Demixer is ready to use.
   bool get isReady => state == DemixerState.done;
 
+  /// Whether the Demixer is loading.
   bool get isLoading => [
         DemixerState.uploading,
         DemixerState.separating,
@@ -118,12 +75,12 @@ class Demixer extends MusicPlayerComponent {
 
     loadingStateNotifier.value = DemixerState.uploading;
 
-    UploadResponse response = await _api.uploadYoutubeSong(song.id);
+    UploadResponse response = await api.uploadYoutubeSong(song.id);
     String songName = response.songName;
 
     if (response.jobId != null) {
       loadingStateNotifier.value = DemixerState.separating;
-      var subscription = _api.jobProgress(response.jobId!).handleError((error) {
+      var subscription = api.jobProgress(response.jobId!).handleError((error) {
         print("ERROR: $error");
         if (error is! JobNotFoundException) throw error;
       }).listen((response) {
@@ -233,15 +190,15 @@ class Demixer extends MusicPlayerComponent {
 
   Future<void> onEqualizerChanged() async {
     // TODO: Get this to work. Currently, there is no trigger for when the Equalizer's parameters changes.
-    var musicPlayerBands = MusicPlayer.instance.equalizer.parameters?.bands;
-    if (musicPlayerBands == null) return;
+    // var musicPlayerBands = MusicPlayer.instance.equalizer.parameters?.bands;
+    // if (musicPlayerBands == null) return;
 
-    for (Stem stem in stems) {
-      var bands = (await stem.equalizer.parameters).bands;
-      for (int i = 0; i < bands.length; i++) {
-        await bands[i].setGain(musicPlayerBands[i].gain);
-      }
-    }
+    // for (Stem stem in stems) {
+    //   var bands = (await stem.equalizer.parameters).bands;
+    //   for (int i = 0; i < bands.length; i++) {
+    //     await bands[i].setGain(musicPlayerBands[i].gain);
+    //   }
+    // }
   }
 
   Future<void> onEnabledToggle() async {
@@ -302,9 +259,15 @@ class Demixer extends MusicPlayerComponent {
 
   /// Save settings for a song to a json map.
   ///
-  /// Saves the following key-value pairs (beyond "enabled"):
-  ///  - "pitchSemitones": [double] How much the pitch will be shifted, in semitones..
-  ///  - "speed": [double] The playback speed of the audio, as a fraction.
+  /// Saves the following stems:
+  ///  - "drums"
+  ///  - "bass"
+  ///  - "vocals"
+  ///  - "other"
+  ///
+  /// Each stem contains the following key-value pairs:
+  ///  - "enabled": [bool] Whether this stem is enabled and should be played.
+  ///  - "volume": [double] The volume this stem is played back at. Must be between 0 and 1.
   @override
   Map<String, dynamic> saveSettingsToJson() {
     return {
