+++ lib/music_player/demixer/demixer_api.dart
@@ -2,36 +2,14 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:http/http.dart' as http;
+import 'package:musbx/music_player/demixer/demixer_api_exceptions.dart';
 import 'package:path_provider/path_provider.dart';
 
-class StemNotFoundException implements Exception {
-  final String? msg;
-
-  const StemNotFoundException([this.msg]);
-
-  @override
-  String toString() => msg ?? 'StemNotFoundException';
-}
-
-class JobNotFoundException implements Exception {
-  final String? msg;
-
-  const JobNotFoundException([this.msg]);
-
-  @override
-  String toString() => msg ?? 'JobNotFoundException';
-}
-
-class ServerException implements Exception {
-  final String? msg;
-
-  const ServerException([this.msg]);
-
-  @override
-  String toString() => msg ?? 'ServerException';
-}
-
 class UploadResponse {
+  /// Returned when uploading a song to the server.
+  ///
+  /// If [jobId] is not `null`, the server has begun separating the song.
+  /// Check the job status with [jobProgress] to make sure the separation job has completed before trying to download stems.
   const UploadResponse(this.songName, {this.jobId});
 
   /// The name of the folder where the stems are saved. Used to download the stems.
@@ -44,12 +22,14 @@ class UploadResponse {
 
 /// A response from a source separation stream.
 class SeparationResponse {
+  /// Returned when checking the status of a job.
   const SeparationResponse(this.progress);
 
-  /// The current progress of the separation.
+  /// The current progress of the separation job.
   final int progress;
 }
 
+/// The stems that can be requested from the server.
 enum StemType {
   drums,
   bass,
@@ -61,8 +41,10 @@ class DemixerApi {
   /// The server hosting the Demixer API.
   final String host = "musbx.agardh.se:8080";
 
+  /// The directory where stems are saved.
   Directory? stemDirectory;
 
+  /// Upload a YouTube song to the server.
   Future<UploadResponse> uploadYoutubeSong(String youtubeId) async {
     Uri url = Uri.http(host, "/upload/$youtubeId");
     var response = await http.post(url);
@@ -80,6 +62,9 @@ class DemixerApi {
   }
 
   /// Check the progress of a separation job.
+  ///
+  /// The progress is checked every [checkEvery] seconds until the job can no
+  /// longer be found (it is completed) and a [JobNotFoundException] is thrown.
   Stream<SeparationResponse> jobProgress(
     String jobId, {
     Duration checkEvery = const Duration(seconds: 5),
@@ -104,7 +89,7 @@ class DemixerApi {
     }
   }
 
-  /// Download a [stem] for a [song].
+  /// Download a [stem] for a [song] to the [stemDirectory].
   Future<File?> downloadStem(String song, StemType stem) async {
     Uri url = Uri.http(host, "/stem/$song/${stem.name}");
     var response = await http.get(url);
