+++ lib/music_player/demixer/demixer_api_exceptions.dart
@@ -1,0 +1,26 @@
+class StemNotFoundException implements Exception {
+  final String? msg;
+
+  const StemNotFoundException([this.msg]);
+
+  @override
+  String toString() => msg ?? 'StemNotFoundException';
+}
+
+class JobNotFoundException implements Exception {
+  final String? msg;
+
+  const JobNotFoundException([this.msg]);
+
+  @override
+  String toString() => msg ?? 'JobNotFoundException';
+}
+
+class ServerException implements Exception {
+  final String? msg;
+
+  const ServerException([this.msg]);
+
+  @override
+  String toString() => msg ?? 'ServerException';
+}
